kong/docker/requirements.txt
requests
python-keycloak